+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [1.4.0-beta.2](https://github.com/wireweave/core/compare/v1.4.0-beta.1...v1.4.0-beta.2) (2026-01-17)
+
+### Bug Fixes
+
+* use theme background color for badge text instead of hardcoded white ([f847845](https://github.com/wireweave/core/commit/f8478457496df8d4a9361f6a1cc8cbd5ff9d16ca))
+
+## [1.4.0-beta.1](https://github.com/wireweave/core/compare/v1.4.0-beta.0...v1.4.0-beta.1) (2026-01-17)
+
+### Bug Fixes
+
+* apply background option only to page element, not theme colors ([7312799](https://github.com/wireweave/core/commit/731279946e033a19d7c91f07f030160a91afce47))
+
+## [1.4.0-beta.0](https://github.com/wireweave/core/compare/v1.3.0...v1.4.0-beta.0) (2026-01-17)
+
+### Features
+
+* **renderer:** add background option to RenderOptions ([626e71b](https://github.com/wireweave/core/commit/626e71b11fdd7ae08d10285a46e13badbda73465))
+
 ## [1.3.0](https://github.com/wireweave/core/compare/v1.3.0-beta.0...v1.3.0) (2026-01-17)
 
 ## [1.3.0-beta.0](https://github.com/wireweave/core/compare/v1.2.0-beta.3...v1.3.0-beta.0) (2026-01-17)